style: revise selected style
