+++ client/src/features/editors/list/List.module.scss
@@ -22,7 +22,7 @@
 }
 
 .cursor {
-  box-shadow: 2px 2px 0 $ontime-pink;
+  box-shadow: 0 0 1px 2px $action-blue;
   border-radius: 3px;
 }
